battle complete

+++ MTCG/MTCG/Card.cs
@@ -6,13 +6,13 @@ namespace MTCG
 {
     public abstract class Card
     {
-        public abstract float Damage { get; set; }
+        public abstract double Damage { get; set; }
         public abstract ElementType ElementType { get; }
         public abstract CardType CardType { get;  }
         public abstract MonsterType MonsterType { get; }
-        public abstract float Attack(Card Card);
+        public abstract double Attack(Card Card);
         public string CardId { get; set; }
-        //public abstract float SetDamage(Card Card, string cardid);
+        //public abstract double SetDamage(Card Card, string cardid);
         public DB db = new DB();
     }
 }
