+++ MTCG/MTCG/MonsterCard.cs
@@ -9,23 +9,23 @@ namespace MTCG
         public override ElementType ElementType { get; }
         public override CardType CardType { get; }
         public override MonsterType MonsterType { get; }
-        public override float Damage { get; set; }
+        public override double Damage { get; set; }
 
-        public MonsterCard(ElementType _element, MonsterType _monsterType, float damage)
+        public MonsterCard(ElementType _element, MonsterType _monsterType, double damage)
         {
             ElementType = _element;
             MonsterType = _monsterType;
             CardType = CardType.Monster;
             Damage = damage;
         }
-        /*public override float Attack(Card Card)
+        /*public override double Attack(Card Card)
         {
             return Damage;
         }*/
 
-        public override float Attack(Card Card)
+        public override double Attack(Card Card)
         {
-            float Temp = Damage;
+            double Temp = Damage;
             if (Card.CardType == CardType.Spell)
             {
                 if (ElementType == ElementType.Fire && Card.ElementType == ElementType.Water
