+++ MTCG/MTCG/MonsterCards/Dragon.cs
@@ -3,16 +3,15 @@ using System.Collections.Generic;
 using System.Text;
 
 namespace MTCG.MonsterCards
-{
-    
+{ 
     public class Dragon : MonsterCard
     {
-        public Dragon(float damage) : base (ElementType.Fire, MonsterType.Dragon, damage) 
+        public Dragon(double damage) : base (ElementType.Fire, MonsterType.Dragon, damage) 
         {
 
         }
 
-        public override float Attack(Card Card)
+        public override double Attack(Card Card)
         {
             if(Card.MonsterType == MonsterType.FireElve)
             {
